DAO EJB

+++ src/main/java/com/rizomm/matgot/marieu/fou/model/Product.java
@@ -76,7 +76,7 @@ public class Product implements Serializable {
         return price;
     }
 
-    public void setPrice(Float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
